feat: upload file

+++ src/service/SekolahService.ts
@@ -2,6 +2,7 @@ import { Repository } from "typeorm";
 import { AppDataSource } from "../data-source";
 import { Sekolah } from "../entities/Sekolah";
 import { createSekolahValidation } from "../libs/sekolah-validation";
+import { v2 as cloudinary } from "cloudinary";
 
 class SekolahService {
   private readonly sekolahRepository: Repository<Sekolah> =
@@ -26,6 +27,12 @@ class SekolahService {
           id: id,
         },
       });
+
+      if (!sekolah) {
+        return {
+          message: "Sekolah not found!",
+        };
+      }
       return {
         message: "Successfully Get Sekolah",
         data: sekolah,
@@ -35,40 +42,61 @@ class SekolahService {
     }
   }
 
-  async create(reqBody: any): Promise<any> {
+  async create(reqBody: any, filename?: any): Promise<any> {
     try {
-      const { error } = createSekolahValidation.validate(reqBody);
-      if (error) {
+      // const { error } = createSekolahValidation.validate(reqBody);
+      // if (error) {
+      //   return {
+      //     message: "Validation error",
+      //     error: error.details[0].message,
+      //     status: 400,
+      //   };
+      // }
+
+      if (filename) {
+        const cloudinaryResponse = await cloudinary.uploader.upload(
+          "./uploads/" + filename
+        );
+        const sekolah = this.sekolahRepository.create({
+          nama_sekolah: reqBody.nama_sekolah,
+          nomor_telepon: reqBody.nomor_telepon,
+          email: reqBody.email,
+          website: reqBody.website,
+          alamat: reqBody.alamat,
+          logo: cloudinaryResponse.secure_url,
+          status: reqBody.status,
+        });
+
+        await this.sekolahRepository.save(sekolah);
+
         return {
-          message: "Validation error",
-          error: error.details[0].message,
-          status: 400,
+          message: "Sekolah successfully Created!",
+          data: sekolah,
+        };
+      } else {
+        const sekolah = this.sekolahRepository.create({
+          nama_sekolah: reqBody.nama_sekolah,
+          nomor_telepon: reqBody.nomor_telepon,
+          email: reqBody.email,
+          website: reqBody.website,
+          alamat: reqBody.alamat,
+          status: reqBody.status,
+        });
+
+        await this.sekolahRepository.save(sekolah);
+
+        return {
+          message: "Sekolah successfully Created!",
+          data: sekolah,
         };
       }
-
-      const sekolah = this.sekolahRepository.create({
-        nama_sekolah: reqBody.nama_sekolah,
-        nomor_telepon: reqBody.nomor_telepon,
-        email: reqBody.email,
-        website: reqBody.website,
-        alamat: reqBody.alamat,
-        logo: reqBody.logo,
-        status: reqBody.status,
-      });
-
-      await this.sekolahRepository.save(sekolah);
-
-      return {
-        message: "Sekolah successfully Created!",
-        data: sekolah,
-      };
     } catch (err) {
       console.error("Error in create method:", err);
       throw new Error("Something Wrong with the server");
     }
   }
 
-  async update(reqBody?: any, id?: any): Promise<any> {
+  async update(reqBody?: any, id?: any, filename?: any): Promise<any> {
     try {
       const sekolah = await this.sekolahRepository.findOneOrFail({
         where: {
@@ -76,13 +104,20 @@ class SekolahService {
         },
       });
 
-      sekolah.nama_sekolah = reqBody.nama_sekolah;
-      sekolah.nomor_telepon = reqBody.nomor_telepon;
-      sekolah.email = reqBody.email;
-      sekolah.website = reqBody.website;
-      sekolah.alamat = reqBody.alamat;
-      sekolah.logo = reqBody.logo;
-      sekolah.status = reqBody.status;
+      sekolah.nama_sekolah = reqBody.nama_sekolah || sekolah.nama_sekolah;
+      sekolah.nomor_telepon = reqBody.nomor_telepon || sekolah.nomor_telepon;
+      sekolah.email = reqBody.email || sekolah.email;
+      sekolah.website = reqBody.website || sekolah.website;
+      sekolah.alamat = reqBody.alamat || sekolah.alamat;
+      sekolah.status = reqBody.status || sekolah.status;
+
+      if (filename) {
+        const cloudinaryResponse = await cloudinary.uploader.upload(
+          "./uploads/" + filename
+        );
+
+        sekolah.logo = cloudinaryResponse.secure_url || sekolah.logo;
+      }
 
       const sekolahUpdate = await this.sekolahRepository.save(sekolah);
       return {
@@ -96,6 +131,17 @@ class SekolahService {
 
   async delete(id: any): Promise<any> {
     try {
+      const sekolah = await this.sekolahRepository.findOne({
+        where: {
+          id: id,
+        },
+      });
+
+      if (!sekolah) {
+        return {
+          message: "Sekolah not found!",
+        };
+      }
       await this.sekolahRepository.delete(id);
 
       return {
